// services/backend/src/types/extentCode.ts
export enum ExtentCode {
  BACHELOR = 1,
  MASTER = 2,
  LICENTIATE = 3,
  DOCTOR = 4,
  BACHELOR_AND_MASTER = 5,
  CONTINUING_EDUCATION = 6,
  EXCHANGE_STUDIES = 7,
  OPEN_UNIVERSITY_STUDIES = 9,
  NON_DEGREE_PEGAGOGICAL_STUDIES_FOR_TEACHERS = 13,
  CONTRACT_TRAINING = 14,
  STUDIES_FOR_SECONDARY_SCHOOL_STUDENTS = 16,
  SPECIALIZATION_STUDIES = 18,
  NON_DEGREE_PROGRAMME_FOR_SPECIAL_EDUCATION_TEACHERS = 22,
  SPECIALIST_TRAINING_IN_MEDICINE_AND_DENTISTRY = 23,
  SUMMER_AND_WINTER_SCHOOL = 31,
  EXCHANGE_STUDIES_POSTGRADUATE = 34,
  NON_DEGREE_STUDIES = 99,
}